add class filters &  fixed issue with reset line edit
