+++ Content.Shared/StationRecords/SharedGeneralStationRecordConsoleSystem.cs
@@ -59,37 +59,3 @@ public sealed class SelectGeneralStationRecord : BoundUserInterfaceMessage
         SelectedKey = selectedKey;
     }
 }
-
-[Serializable, NetSerializable]
-public sealed class GeneralStationRecordsFilterMsg : BoundUserInterfaceMessage
-{
-    public string value { get; }
-    public GeneralStationRecordFilterType type { get; }
-
-    public GeneralStationRecordsFilterMsg(GeneralStationRecordFilterType filterType,
-        string  filterValue)
-    {
-        type = filterType;
-        value = filterValue;
-    }
-}
-
-[Serializable, NetSerializable]
-public enum GeneralStationRecordFilterType : byte
-{
-    Name,
-    Prints,
-    DNA,
-}
-
-[Serializable, NetSerializable]
-public sealed class GeneralStationRecordsFilter {
-    public GeneralStationRecordFilterType type { get; set; } =
-        GeneralStationRecordFilterType.Name;
-    public string value { get; set; } = "";
-    public GeneralStationRecordsFilter(GeneralStationRecordFilterType filterType,
-        string newValue = "") {
-            type = filterType;
-            value = newValue;
-    }
-}
